+++ R/tbl.R
@@ -85,7 +85,9 @@ group_by.tbl_kusto_abstract <- function(.data, ..., add = FALSE)
         return(.data)
     }
 
-    groups <- group_by_prepare(.data, .dots = dots, add = add)
+    groups <- group_by_prepare(.data, .dots = dots, .add = add)
+    # TODO: .dots is deprecated
+    #groups <- group_by_prepare(.data, across(dots), .add = add)
     names <- vapply(groups$groups, as_string, character(1))
     add_op_single("group_by",
                   groups$data,
@@ -135,45 +137,6 @@ summarise.tbl_kusto_abstract <- function(.data, ..., .strategy = NULL, .shufflek
     dots <- partial_eval(dots, vars = op_vars(.data))
     add_op_single("summarise", .data, dots = dots,
                   args = list(.strategy = .strategy, .shufflekeys = .shufflekeys, .num_partitions = .num_partitions))
-}
-
-#' Unnest method for Kusto tables
-#'
-#' This method takes a list column and expands it so that each element of the list gets its own row.
-#' unnest() translates to Kusto's mv-expand operator.
-#'
-#' @param .data A Kusto tbl.
-#' @param ... Specification of columns to unnest.
-#' @param .id Data frame identifier - if supplied, will create a new column with name .id, giving a unique identifier. This is most useful if the list column is named.
-#' @export
-unnest.tbl_kusto_abstract <- function(.data, ..., .id = NULL)
-{
-    dots <- quos(...)
-    add_op_single("unnest", .data, dots = dots, args = list(.id = .id))
-}
-
-#' Nest method for Kusto tables
-#'
-#' This method collapses a column into a list
-#'
-#' @param .data A kusto tbl.
-#' @param ... Specification of columns to nest. Translates to summarize make_list() in Kusto.
-#' @export
-nest.tbl_kusto_abstract <- function(.data, ...)
-{
-    nest_vars <- unname(tidyselect::vars_select(op_vars(.data), ...))
-
-    if (is_empty(nest_vars))
-        nest_vars <- op_vars(.data)
-
-    group_vars <- union(op_grps(.data), setdiff(op_vars(.data), nest_vars))
-    nest_vars <- setdiff(nest_vars, group_vars)
-    dot_calls <- mapply(function(x) expr(make_list(!!as.name(x))), nest_vars)
-
-    if (is_empty(group_vars))
-        summarise(.data, !!! dot_calls)
-    else
-        summarise(group_by(.data, !! as.name(group_vars)), !!! dot_calls)
 }
 
 #' @export