+++ tests/testthat/test02_query.R
@@ -20,7 +20,6 @@ server <- sprintf("https://%s.%s.kusto.windows.net", srvname, srvloc)
 db <- kusto_database_endpoint(server=server, database=dbname, tenantid=tenant)
 db2 <- kusto_database_endpoint(server=server, database=dbname, tenantid=tenant,
     appclientid=app, appkey=password)
-db3 <- kusto_database_endpoint(server=server, database=dbname)
 
 test_that("Queries work",
 {
@@ -58,26 +57,6 @@ test_that("Commands work with own app",
     expect_is(out, "data.frame")
 
     dberr <- db2
-    dberr$token <- NULL
-    expect_error(run_query(dberr, ".show cluster"))
-})
-
-test_that("Queries work with common tenant",
-{
-    out <- run_query(db3, "iris | summarize count() by species")
-    expect_is(out, "data.frame")
-
-    dberr <- db3
-    dberr$token <- NULL
-    expect_error(run_query(dberr, "iris | summarize count() by species"))
-})
-
-test_that("Commands work with common tenant",
-{
-    out <- run_query(db3, ".show cluster")
-    expect_is(out, "data.frame")
-
-    dberr <- db3
     dberr$token <- NULL
     expect_error(run_query(dberr, ".show cluster"))
 })